integrate pundit with active_interaction. finished up user interactions.

+++ app/interactions/users/destroy_user.rb
@@ -1,7 +1,12 @@
-class DestroyUser < ActiveInteraction::Base
+class DestroyUser < ApplicationInteraction
+  object :current_user, class: User
   object :user
 
   def execute
+    authorize(current_user, user, :destroy?)
+
+    return user unless errors.messages.empty?
+
     user.destroy
 
     user
